tools/requirements.txt
httpx>=0.27.0
rich>=13.0.0
python-dotenv>=1.0.0
python-slugify>=8.0.0
